Move DUA from access.py to dua.py.
